refactor: renaming error utility
